bug fix more than one text/plain mime part
i had an issue where an email message had 2 text/plain mime parts. line 193 then executes on both of them.
in my case the 2nd mime part had no TextContent and it overwrote the correct BodyText.Text setting it to empty string.
I added a small fix to check if the TextContent part is not empty in the same if statement. Thanks!